Reverted converting log_num_frames to int as it is being multiplied by a float type mask

+++ babyai/rl/algos/base.py
@@ -101,12 +101,12 @@ class BaseAlgo(ABC):
 
         self.log_episode_return = torch.zeros(self.num_procs, device=self.device)
         self.log_episode_reshaped_return = torch.zeros(self.num_procs, device=self.device)
-        self.log_episode_num_frames = torch.zeros(self.num_procs, device=self.device, dtype=torch.int)
+        self.log_episode_num_frames = torch.zeros(self.num_procs, device=self.device)
 
         self.log_done_counter = 0
         self.log_return = [0.0] * self.num_procs
         self.log_reshaped_return = [0.0] * self.num_procs
-        self.log_num_frames = [0] * self.num_procs
+        self.log_num_frames = [0.0] * self.num_procs
 
     def collect_experiences(self): #note that we are collecting experiences with torch.no_grad!
         """Collects rollouts and computes advantages.
@@ -185,7 +185,7 @@ class BaseAlgo(ABC):
 
             self.log_episode_return += torch.tensor(reward, device=self.device, dtype=torch.float)
             self.log_episode_reshaped_return += self.rewards[i]
-            self.log_episode_num_frames += torch.ones(self.num_procs, device=self.device, dtype=torch.int)
+            self.log_episode_num_frames += torch.ones(self.num_procs, device=self.device)
 
             for i, done_ in enumerate(done):
                 if done_: # the foll'ing get updated/logged only when done in curr step, else they are all zero per proc
@@ -260,10 +260,10 @@ class BaseAlgo(ABC):
             "return_per_episode": torch.tensor(self.log_return[-keep:], device=self.device), # list of rewards when done
             # happens within an episode; like avg return within an episode (when the list is avg'd)
             "reshaped_return_per_episode": torch.tensor(self.log_reshaped_return[-keep:], device=self.device),
-            "num_frames_per_episode": torch.tensor(self.log_num_frames[-keep:], device=self.device, dtype=torch.int),
+            "num_frames_per_episode": torch.tensor(self.log_num_frames[-keep:], device=self.device),
             # avg num of # frames taken for 'done'
             "num_frames": self.num_frames,
-            "episodes_done": torch.tensor(self.log_done_counter, device=self.device, dtype=torch.int)
+            "episodes_done": torch.tensor(self.log_done_counter, device=self.device)
         }
 
         self.log_done_counter = 0
